added resize functionality

+++ wio.py
@@ -33,7 +33,8 @@ WELCOME_MESSAGE = f"""__        _____ ___
   \ V  V /  | | |_| |
    \_/\_/  |___\___/ 
 Web Image Optimizer
-v. {VERSION}
+Made with ❤️ by UndyingSoul
+GitHub: https://github.com/UndyingSoul/wio
     """
 
 # Supported image formats and their corresponding Pillow plugins
@@ -49,6 +50,8 @@ SUPPORTED_FORMATS = {
     "avif": AvifImagePlugin.AvifImageFile,
 }
 
+human_readable_bools = ('no','yes')
+
 def range_arg_type(astr, min=0, max=100):
     value = int(astr)
     if min <= value <= max:
@@ -63,17 +66,41 @@ def size_of_file_fmt(num, suffix="B"):
         num /= 1024.0
     return f"{num:.1f} Yi{suffix}"
 
+def generate_output_path(input_path, format, keep_originals, save_as_original_format):
+    directory, filename = os.path.split(input_path)
+    filename, ext = os.path.splitext(filename)
+    
+    if save_as_original_format:
+        new_ext = ext if not keep_originals else "_wio" + ext
+    else:
+        new_ext = + ext
+
+    output_path = os.path.join(directory, filename + new_ext)
+    return output_path
+
 # Function to convert an image to the specified format and quality
-async def convert_image(input_path, output_path, format, quality, keep_metadata):
+async def convert_image(input_path, output_path, format, quality, keep_metadata, keep_originals, save_as_original_format, resize=False):
     log.debug(f"Converting image {input_path}")
     try:
         with Image.open(input_path) as img:
             img = ImageOps.exif_transpose(img)  # Fix orientation if applicable
             
+            if resize:
+                width, height = img.size
+                new_width = int(width * (quality / 100))
+                new_height = int(height * (quality / 100))
+                img = img.resize((new_width, new_height), Image.LANCZOS)
+
             exif_data = img.info.get('exif')  # Get exif data
             exif_kwargs = {'exif': exif_data} if exif_data and keep_metadata else {}
 
-            img.save(output_path, format=format, quality=quality, **exif_kwargs)
+            if save_as_original_format:
+                output_path = generate_output_path(input_path, format, keep_originals, save_as_original_format)
+
+            if resize:
+                img.save(output_path, format=format, quality=95, subsampling=0, **exif_kwargs)
+            else:
+                img.save(output_path, format=format, quality=quality, **exif_kwargs)
             return True
     except Exception as e:
         log.error(f"Could not convert {input_path}")
@@ -81,16 +108,31 @@ async def convert_image(input_path, output_path, format, quality, keep_metadata)
         return False
 
 # Function to process a single image file
-async def process_file(file_path, format, quality, keep_originals, keep_metadata, index, total):
+async def process_file(file_path, format, quality, keep_originals, keep_metadata, index, total, resize=False):
     directory, filename = os.path.split(file_path)
     filename, ext = os.path.splitext(filename)
-    output_path = os.path.join(directory, filename + "." + format)
-    if await convert_image(file_path, output_path, format, quality, keep_metadata):
-        file_path_size = os.stat(file_path).st_size
+    save_as_original_format = False
+    
+    if format == "original":
+        save_as_original_format = True
+        format = ext[1:]
+        if format == "jpg":
+            format = "jpeg" # dumb, but necessary edge case
+            
+        
+        output_path = generate_output_path(file_path, format, keep_originals, save_as_original_format)
+        if not format:
+            log.critical(f"Invalid extension for file {output_path}")
+            return
+    else:
+        output_path = os.path.join(directory, filename + "." + format)
+        
+    file_path_size = os.stat(file_path).st_size
+    if await convert_image(file_path, output_path, format, quality, keep_metadata, keep_originals, save_as_original_format, resize):
         output_path_size = os.stat(output_path).st_size
         space_saved_percent = int((1-(output_path_size/file_path_size))*100)
-        space_saved = f"({space_saved_percent}% smaller)" if space_saved_percent > 0 else f"{Fore.BLACK}{Back.WHITE}({abs(space_saved_percent)}% larger){Fore.RESET}{Back.RESET}"
-        if not keep_originals:  # Optionally delete the original file
+        space_saved = f"({space_saved_percent}% smaller)" if space_saved_percent > -1 else f"{Fore.BLACK}{Back.WHITE}({abs(space_saved_percent)}% larger){Fore.RESET}{Back.RESET}"
+        if not keep_originals and not save_as_original_format:  # Optionally delete the original file
             log.info(f"Deleting file {file_path}")
             os.remove(file_path)
         log.info(f"{index+1}/{total} ({int(((index+1)/total)*100)}%) - Converted {file_path}[{size_of_file_fmt(file_path_size)}] to {output_path}[{size_of_file_fmt(output_path_size)}] {space_saved}")
@@ -132,6 +174,12 @@ def main():
         metavar="[0-100]",
         default=70,
         help="Quality of the conversion (0-100)",
+    )
+    parser.add_argument(
+        "--resize",
+        "-r",
+        action="store_true",
+        help="Resize images instead of optimizing them",
     )
     parser.add_argument(
         "--keep-originals",
@@ -176,12 +224,19 @@ def main():
         elif os.path.isfile(path) and any(path.lower().endswith(ext) for ext in args.extensions):
             log.debug(f"{path} is file")
             file_list.append(path)
+        else:
+            log.error(f"Unsupported image or directory {path}")
+
+    if not file_list:
+        log.error("No valid files specified. Exiting.")
+        raise SystemExit(0)
     
     log.info(f"Converting {len(file_list)} files to the following parameters\n" + 
              f" * Format: {args.format}\n" + 
              f" * Quality: {args.quality}%\n" +
-             f" * Keeping Originals: {args.keep_originals}\n" + 
-             f" * Keeping Metadata: {args.keep_metadata}")
+             f" * Resize Images: {human_readable_bools[args.resize]}\n" +
+             f" * Keeping Originals: {human_readable_bools[args.keep_originals]}\n" + 
+             f" * Keeping Metadata: {human_readable_bools[args.keep_metadata]}")
     
     log.log(logging.HEADER, "File Conversion")
 
@@ -191,7 +246,7 @@ def main():
         if args.dry_run:
             log.info(f"Will convert {file_path}")
             continue
-        asyncio.run(process_file(file_path, args.format, args.quality, args.keep_originals, args.keep_metadata, index, len(file_list)))
+        asyncio.run(process_file(file_path, args.format, args.quality, args.keep_originals, args.keep_metadata, index, len(file_list), resize=args.resize))
     
 
     seconds = time.time() - start_time
